restrict ports to 0 -> 2^16 - 1

+++ src/server_test.go
@@ -17,7 +17,7 @@ func TestNewServerDefaults(t *testing.T) {
 
 func TestNewServerOnlyPort(t *testing.T) {
 	srv := NewServer(ServerConfig{Port: 3000})
-	wantPort := 3000
+	wantPort := uint16(3000)
 
 	if srv.conf.Port != wantPort {
 		t.Errorf(`custom port = %d, want %d`, srv.conf.Port, wantPort)
